+++ api/common/src/auth/SecretManager.js
